refactor: move MyQPaxos impl into standalone dir
