+++ components/epaxos/src/serverdata/errors.rs
@@ -1,5 +1,3 @@
-
-extern crate parse;
 use parse::Response;
 
 quick_error! {
