+++ components/epaxos/src/service/mod.rs
@@ -1,0 +1,2 @@
+mod service;
+pub use service::*;
